services/vectorstore/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pydantic==2.5.0
psutil
sentence-transformers>=2.5.1
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2
torch>=2.6.0
transformers==4.54.1 